commit before pull [conflict only in onekjagay]

+++ app/src/main/java/com/example/covid_19alertapp/roomdatabase/VisitedLocations.java
@@ -12,7 +12,7 @@ public class VisitedLocations {
     @NonNull
     private String conatainerDateTimeComposite;
 
-    private int count;
+    private long count;
 
     public VisitedLocations() {
         /*
@@ -21,7 +21,7 @@ public class VisitedLocations {
     }
 
     @Ignore
-    public VisitedLocations(String conatainerDateTimeComposite, int count) {
+    public VisitedLocations(String conatainerDateTimeComposite, long count) {
         this.conatainerDateTimeComposite = conatainerDateTimeComposite;
         this.count = count;
     }
@@ -38,7 +38,7 @@ public class VisitedLocations {
         this.conatainerDateTimeComposite = conatainerDateTimeComposite;
     }
 
-    public void setCount(int count) {
+    public void setCount(long count) {
         this.count = count;
     }
 
@@ -46,7 +46,7 @@ public class VisitedLocations {
         return conatainerDateTimeComposite;
     }
 
-    public int getCount() {
+    public long getCount() {
         return count;
     }
